Refactor immersive editor logic into a composable and move shared styles
